third_party/blink/web_tests/http/tests/websocket/websocket-event-target-expected.txt

Make sure WebSocket object acts as EventTarget.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS open_event_handled is false
PASS message_event_handled is false
PASS close_event_handled is false
PASS open_event_handled is true
PASS open_event_handled is false
PASS message_event_handled is true
PASS message_event_handled is false
PASS close_event_handled is true
PASS close_event_handled is false
PASS successfullyParsed is true

TEST COMPLETE
